chore: moved ContextMessage into its own JS file
