+++ Sebedius.js
@@ -4,6 +4,7 @@ const Discord = require('discord.js');
 const YZCrit = require('./yearzero/YZCrit');
 const Util = require('./utils/Util');
 const PageMenu = require('./utils/PageMenu');
+const ContextMessage = require('./utils/ContextMessage');
 const RollTable = require('./utils/RollTable');
 const Errors = require('./utils/errors');
 const { SUPPORTED_GAMES, DICE_ICONS, SOURCE_MAP } = require('./utils/constants');
@@ -522,7 +523,7 @@ class Sebedius extends Discord.Client {
 				embed.setFooter(`page ${page + 1}/${paginatedChoices.length}`);
 			}
 			if (text) {
-				embed.addField('Note', text, false);
+				embed.addField('Info', text, false);
 			}
 			if (pm) {
 				embed.addField(
@@ -717,45 +718,3 @@ function whenMentionedOrPrefixed(prefixes, client) {
 }
 
 module.exports = Sebedius;
-
-/**
- * Represents a Discord message with context.
- * @extends {Discord.Message}
- * @see Sebedius.processMessage
- */
-class ContextMessage extends Discord.Message {
-	/**
-	 * @param {string} prefix The prefix used to trigger the command
-	 * @param {Discord.Client} client The instantiating client
-	 * @param {Object} data The data for the message
-	 * @param {Discord.TextChannel|Discord.DMChannel|Discord.NewsChannel} channel The channel the message was sent in
-	 */
-	constructor(prefix, client, data, channel) {
-		super(client, data, channel);
-
-		/**
-		 * The prefix used to trigger the command.
-		 * @type {string}
-		 */
-		this.prefix = prefix;
-	}
-
-	/**
-	 * The bot client (Sebedius).
-	 * @type {Discord.Client}
-	 * @readonly
-	 */
-	get bot() { return this.client; }
-
-	/**
-	 * Sends a message to the channel.
-	 * @param {StringResolvable|Discord.APIMessage} [content=''] The content to send
-	 * @param {Discord.MessageOptions|Discord.MessageAdditions} [options={}] The options to provide
-	 * @returns {Promise<Discord.Message|Discord.Message[]>}
-	 * @async
-	 */
-	async send(content, options) {
-		// if (this.channel.type === 'dm') return await this.author.send(content, options);
-		return await this.channel.send(content, options);
-	}
-}
